refactor: update folder structure
