--- src/components/CheckboxCell/styles.css
@@ -1,4 +1,0 @@
-.NLT__checkbox-cell {
-    width: 100%;
-    padding: var(--nlt-cell-spacing);
-}
