--- src/components/DateCellEdit/styles.css
@@ -1,5 +1,0 @@
-.NLT__date-input {
-    width: 100%;
-    padding: 0px 10px !important;
-    border: 0 !important;
-}
