--- src/components/EditableTd/styles.css
@@ -1,8 +1,0 @@
-.NLT__td-container {
-	display: flex;
-	width: 100%;
-	min-height: var(--nlt-cell-min-height);
-	height: 100%;
-	padding: var(--nlt-cell-spacing-x) var(--nlt-cell-spacing-y);
-	cursor: pointer;
-}
