--- src/components/EditableTh/styles.css
@@ -1,30 +1,0 @@
-.NLT__th-container {
-    display: flex;
-    justify-content: space-between;
-    width: 100%;
-    min-height: var(--nlt-cell-min-height);
-}
-
-.NLT__th-content {
-    display: flex;
-    align-items: center;
-    /* Use 100% so that the resize indicator appears at the end */
-    width: 100%;
-    overflow: hidden;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-    user-select: none;
-    padding: var(--nlt-cell-spacing-x) var(--nlt-cell-spacing-y);
-}
-
-.NLT__th-resize-container {
-    position: relative;
-}
-
-.NLT__th-resize {
-    position: absolute;
-    left: -10px;
-    cursor: col-resize;
-    width: 10px;
-    height: 100%;
-}
