--- src/components/HeaderMenu/styles.css
@@ -1,12 +1,0 @@
-.NLT__header-menu {
-	color: var(--text-normal);
-}
-
-.NLT__header-menu-title-container {
-	display: flex;
-	font-size: 0.95rem;
-}
-
-.NLT__header-menu-title {
-	font-weight: 700;
-}
