--- src/components/MultiTagCell/styles.css
@@ -1,4 +1,0 @@
-.NLT__multi-tag-cell {
-    display: flex;
-    flex-direction: column;
-}
