--- src/components/NumberCell/styles.css
@@ -1,4 +1,0 @@
-.NLT__number-cell {
-    width: 100%;
-    text-align: right;
-}
