--- src/components/NumberCellEdit/styles.css
@@ -1,8 +1,0 @@
-.NLT__number-cell-edit {
-    width: 100%;
-    height: 100%;
-    text-align: right;
-    border: 0 !important;
-    font-family: var(--font-text);
-    font-size: 1rem;
-}
