--- src/components/TagCell/styles.css
@@ -1,3 +1,0 @@
-.NLT__tag-cell {
-    width: 100%;
-}
