--- src/components/TagCellEdit/component/CreateTag/styles.css
@@ -1,7 +1,0 @@
-.NLT__create-tag {
-    display: flex;
-    align-items: center;
-    padding: 4px 6px;
-    width: 100%;
-    overflow: hidden;
-}
