--- src/components/TagCellEdit/styles.css
@@ -1,17 +1,0 @@
-.NLT__tag-menu {
-	background-color: var(--background-primary);
-}
-
-.NLT__tag-menu-text {
-	font-size: 0.85rem;
-	margin: 5px 0px;
-	width: 100%;
-}
-
-.NLT__tag-input {
-	background-color: transparent !important;
-	border: 0 !important;
-	width: 100%;
-	padding-left: 5px !important;
-	padding-right: 5px !important;
-}
