--- src/components/TagColorMenu/components/ColorItem/styles.css
@@ -1,14 +1,0 @@
-.NLT__color-item {
-    display: flex;
-    align-items: center;
-    font-size: 0.8rem;
-    padding: 0px 5px;
-    width: 100%;
-}
-
-.NLT__color-item-square {
-    width: 10px;
-    height: 10px;
-    padding: 0;
-    margin: 0 10px 0 0;
-}
