--- src/components/TagColorMenu/styles.css
@@ -1,3 +1,0 @@
-.NLT__tag-color-menu {
-    padding: 4px 10px;
-}
