--- src/components/TextCell/styles.css
@@ -1,7 +1,0 @@
-.NLT__text-cell {
-    width: 100%;
-}
-
-.NLT__text-cell p {
-    margin: 0;
-}
