--- src/components/TextCellEdit/styles.css
@@ -1,8 +1,0 @@
-.NLT__textarea {
-    width: 100%;
-    height: 100%;
-    border: 0;
-    overflow: hidden;
-    padding: var(--nlt-cell-spacing-x) var(--nlt-cell-spacing-y);
-    resize: none;
-}
